update wordpress dockerfile and script configs

+++ srcs/requirements/wordpress/tools/wp_script.sh
@@ -1,26 +1,35 @@
-#!/bin/sh
-
-#WP-CLI is available as a PHP Archive file (.phar). You can download it using either wget or curl commands:
-wget https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-
-#You need to make this .phar file executable and move it to /usr/local/bin so that it can be run directly:
-chmod 777 wp-cli.phar
-
-mv wp-cli.phar /usr/local/bin/wp
-
-#WP-CLI includes a command to download WordPress
-wp core download --allow-root
-
-# wp config create --allow-root --dbname=$DB_NAME --dbuser=$DB_USER --dbpass=$DB_ROOT_PASS --locale=ro_RO
-
-# mv /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
-
-#set up the database credentials for our installation
-wp --allow-root core config --dbhost=mariadb --dbname=wordpress --dbuser=fech-cha --dbpass=root
-chmod 777 /var/www/html/wp-config.php
+#! /bin/sh
 
 
-#install WordPress now, we need to run one last command, while configuring WP-CLI credentials
-wp --allow-root core install --url=fech-cha.42.fr --title=inception --admin_user=fech-cha --admin_password=user --admin_email=farouk.echcharef@usmba.ac.ma
+if [ ! -f /var/www/html/wp-config.php ]; then
+    #WP-CLI is available as a PHP Archive file (.phar). You can download it using either wget or curl commands:
+    wget https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+
+    #You need to make this .phar file executable and move it to /usr/local/bin so that it can be run directly:
+    chmod 777 wp-cli.phar
+    mv wp-cli.phar /usr/local/bin/wp
+
+    cd /var/www/html
+
+    #WP-CLI includes a command to download WordPress
+    wp --allow-root core download 
+    cp wp-config-sample.php wp-config.php
+
+
+    wp config set DB_NAME $DB_NAME --allow-root
+	wp config set DB_USER $DB_USER --allow-root
+	wp config set DB_PASSWORD $DB_USER_PASS --allow-root
+	wp config set DB_HOST $DB_HOST --allow-root
+
+    #set up the database credentials for our installation
+    # wp --allow-root core config --dbhost=$DB_HOST --dbname=$DB_NAME --dbuser=$DB_USER --dbpass=$DB_ROOT_PASS
+
+    chmod 777 /var/www/html/wp-config.php
+
+    #install WordPress now, we need to run one last command, while configuring WP-CLI credentials
+    wp --allow-root core install --url=$DOMAIN --title=$WP_TITLE --admin_user=$WP_USER --admin_password=$WP_PASS --admin_email=$WP_EMAIL
+    wp user create $USR $USER_EMAIL --user_pass=$USER_PASSWORD --allow-root
+
+fi
 
 exec php-fpm7.3  -F -R
